feat: Copy weather arts and icons from mobile to wear application
One of the rubric critereas is a "graphic that summarizes the day's
weatherphic that summarizes the day's weather".

Thus the graphics originally by the mobile app were copied to the
wearable one. The colored arts were rescaled to a smaller size so they
would match the grey icons, being respectively used on interactive and
ambient modes.